choreL routine package installation
