+++ package.json
@@ -21,11 +21,12 @@
     "discord.js": "^14.16.3",
     "dotenv": "^16.4.7",
     "express": "^4.21.0",
+    "firebase-admin": "^13.0.2",
     "https": "^1.0.0",
     "mongodb": "^6.11.0",
     "mqtt": "^5.10.3",
     "node-fetch": "^3.3.2",
-    "nodemon": "^2.0.15",
+    "nodemon": "^2.0.22",
     "open": "^10.1.0",
     "puppeteer": "^23.11.1",
     "textmagic-rest-client": "^1.0.11",
